Change Update to contain both old and new indexes
Since when doing updates to a UICollectionView or UITableView, the "reload" methods expect the "old" indexPaths, and the Diff API was returning the "new" IndexPaths.  So now the Diff API returns both old and new indexes to create IndexPaths with.

The ListUpdate follows the diff change and returns old and new IndexPaths for ".update"